better file setup
