+++ apps/dotcom/sync-worker/src/UserDataSyncer.ts
@@ -16,7 +16,7 @@ import { createSentry } from '@tldraw/worker-shared'
 import { CompiledQuery, Kysely, sql } from 'kysely'
 import throttle from 'lodash.throttle'
 import { Logger } from './Logger'
-import { fetchEverythingSql, parseResultRow } from './fetchEverythingSql'
+import { fetchEverythingSql, parseResultRow } from './fetchEverythingSql.snap'
 import { Environment, TLUserDurableObjectEvent, getUserDoSnapshotKey } from './types'
 import { getReplicator, getStatsDurableObjct } from './utils/durableObjects'
 import { retryOnConnectionFailure } from './utils/retryOnConnectionFailure'
