+++ apps/dotcom/sync-worker/src/fetchEverythingSql.snap.ts
